gpu: nvgpu: runlist MISRA fixes for Rule 10.3
Using u32 with bitops like for_each_set_bit results in MISRA
violation as bitops internally uses unsigned long.
Define tsgid as unsigned long an use (u32) cast when needed.

Jira NVGPU-3379

Reviewed-on: https://git-master.nvidia.com/r/2109680
GVS: Gerrit_Virtual_Submit
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>

+++ drivers/gpu/nvgpu/common/fifo/runlist.c
@@ -134,7 +134,7 @@ static u32 nvgpu_runlist_append_prio(struct fifo_gk20a *f,
 	nvgpu_log_fn(f->g, " ");
 
 	for_each_set_bit(tsgid, runlist->active_tsgs, f->num_channels) {
-		struct tsg_gk20a *tsg = nvgpu_tsg_get_from_id(f->g, tsgid);
+		struct tsg_gk20a *tsg = nvgpu_tsg_get_from_id(f->g, (u32)tsgid);
 		u32 entries;
 
 		if (tsg->interleave_level == interleave_level) {
@@ -177,7 +177,7 @@ static u32 nvgpu_runlist_append_med(struct fifo_gk20a *f,
 	nvgpu_log_fn(f->g, " ");
 
 	for_each_set_bit(tsgid, runlist->active_tsgs, f->num_channels) {
-		struct tsg_gk20a *tsg = nvgpu_tsg_get_from_id(f->g, tsgid);
+		struct tsg_gk20a *tsg = nvgpu_tsg_get_from_id(f->g, (u32)tsgid);
 		u32 entries;
 
 		if (tsg->interleave_level !=
@@ -216,7 +216,7 @@ static u32 nvgpu_runlist_append_low(struct fifo_gk20a *f,
 	nvgpu_log_fn(f->g, " ");
 
 	for_each_set_bit(tsgid, runlist->active_tsgs, f->num_channels) {
-		struct tsg_gk20a *tsg = nvgpu_tsg_get_from_id(f->g, tsgid);
+		struct tsg_gk20a *tsg = nvgpu_tsg_get_from_id(f->g, (u32)tsgid);
 		u32 entries;
 
 		if (tsg->interleave_level !=
